% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cogFunctionalCategories.R
\name{cogFunctionalCategories}
\alias{cogFunctionalCategories}
\title{COG functional categories}
\usage{
cogFunctionalCategories()
}
\value{
\code{DataFrame}.
Contains columns \code{"letter"}, \code{"description"}, and \code{"class"}.
}
\description{
Letter code mappings of COG functional categories.
}
\note{
Updated 2022-06-08.
}
\examples{
object <- cogFunctionalCategories()
print(object)
}
\seealso{
\itemize{
\item http://eggnog5.embl.de/download/eggnog_4.5/COG_functional_categories.txt
}
}
