% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerators.R
\name{EggNOG}
\alias{EggNOG}
\title{EggNOG database annotations}
\usage{
EggNOG(release = c("5.0", "4.5", "4.1", "4.0"))
}
\arguments{
\item{release}{\code{character(1)}.
EggNOG release version (e.g. "4.5").
Currently supported: "4.5", "4.1".
Support for EggNOG 5 will be added in next release.}
}
\value{
\code{EggNOG}.
}
\description{
\href{http://eggnogdb.embl.de}{EggNOG} is a database of biological information
hosted by the EMBL. It is based on the original idea of COGs (\strong{c}lusters of
\strong{o}rthologous \strong{g}roups) and expands that idea to non-supervised
orthologous groups constructed from numerous organisms. eggNOG stands for
\strong{e}volutionary \strong{g}enealogy of \strong{g}enes: \strong{N}on-supervised
\strong{O}rthologous \strong{G}roups.
}
\note{
Updated 2022-06-07.
}
\examples{
x <- EggNOG()
print(x)
}
