% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGlobals.R
\docType{data}
\name{EggNOGTestsURL}
\alias{EggNOGTestsURL}
\title{EggNOG test data URL}
\format{
An object of class \code{character} of length 1.
}
\usage{
EggNOGTestsURL
}
\description{
EggNOG test data URL
}
\note{
Updated 2021-02-21.
}
\examples{
EggNOGTestsURL
}
\keyword{internal}
