% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerators.R
\name{EggNOG}
\alias{EggNOG}
\title{EggNOG database annotations}
\usage{
EggNOG(release = "4.5")
}
\arguments{
\item{release}{\code{character} or \code{NULL}.
EggNOG release version (e.g. "5.0").
If set \code{NULL}, will download the latest release.}
}
\value{
\code{EggNOG}.
}
\description{
\href{http://eggnogdb.embl.de}{EggNOG} is a database of biological information
hosted by the EMBL. It is based on the original idea of COGs (\strong{c}lusters of
\strong{o}rthologous \strong{g}roups) and expands that idea to non-supervised
orthologous groups constructed from numerous organisms. eggNOG stands for
\strong{e}volutionary \strong{g}enealogy of \strong{g}enes: \strong{N}on-supervised
\strong{O}rthologous \strong{G}roups.
}
\note{
Updated 2021-02-21.
}
\examples{
x <- EggNOG()
print(x)
}
